improvements

+++ src/functions/todos.ts
@@ -4,45 +4,66 @@ import { prisma } from "@/lib/prisma";
 import { revalidatePath } from "next/cache";
 
 export async function createTodos(title: string, id: string) {
-  await prisma.todo.create({
-    data: {
-      title,
-      userId: id,
-    },
-  });
+  try {
+    await prisma.todo.create({
+      data: {
+        title,
+        userId: id,
+      },
+    });
 
-  revalidatePath("/");
+    revalidatePath("/");
+  } catch (error) {
+    console.log(error);
+  }
 }
 
 export async function deleteTodos(id: string) {
-  await prisma.todo.delete({
-    where: {
-      id,
-    },
-  });
+  // await prisma.todo.deleteMany();
+  // return await prisma.user.deleteMany();
+  try {
+    const todoExists = await prisma.todo.findUnique({ where: { id } });
+    if (!todoExists) return;
 
-  revalidatePath("/");
+    await prisma.todo.delete({
+      where: {
+        id,
+      },
+    });
+
+    revalidatePath("/");
+  } catch (error) {
+    console.log(error);
+  }
 }
 
 export async function setCompletedTodo(id: string, completed: boolean) {
-  await prisma.todo.update({
-    where: {
-      id,
-    },
-    data: {
-      completed,
-    },
-  });
+  try {
+    await prisma.todo.update({
+      where: {
+        id,
+      },
+      data: {
+        completed,
+      },
+    });
 
-  revalidatePath("/");
+    revalidatePath("/");
+  } catch (error) {
+    console.log(error);
+  }
 }
 
 export async function updateTodos(id: string, title: string) {
-  const result = await prisma.todo.update({
-    where: { id },
-    data: { title },
-  });
-  console.log(result);
+  try {
+    const result = await prisma.todo.update({
+      where: { id },
+      data: { title },
+    });
+    console.log(result);
 
-  revalidatePath("/");
+    revalidatePath("/");
+  } catch (error) {
+    console.log(error);
+  }
 }
